+++ src/main/java/com/artyom/romashkako/product/model/Product.java
@@ -2,9 +2,6 @@ package com.artyom.romashkako.product.model;
 
 import lombok.*;
 
-import java.math.BigDecimal;
-import java.util.Comparator;
-
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
@@ -15,7 +12,7 @@ public class Product implements Comparable<Product> {
     private Integer id;
     private String title;
     private String description;
-    private BigDecimal price;
+    private double price;
     private boolean isAvailable;
 
     @Override
