+++ src/test/java/com/artyom/romashkako/product/service/DefaultProductServiceTest.java
@@ -19,12 +19,12 @@ import static org.mockito.Mockito.doThrow;
 class DefaultProductServiceTest {
 
     private static final List<Product> PRODUCT_LIST = new ArrayList<>(List.of(
-            new Product(1, "Роза", "Красный цветок с шипами", BigDecimal.valueOf(34.99), true),
-            new Product(2, "Тюльпан", "Весенний цветок различных оттенков", BigDecimal.valueOf(20.99), true),
-            new Product(3, "Гвоздика", "Цветок с гофрированными лепестками", BigDecimal.valueOf(20.99), false),
-            new Product(4, "Пионы", "Крупный цветок с пышными лепестками", BigDecimal.valueOf(29.99), true),
-            new Product(5, "Лилии", "Элегантный цветок с сильным ароматом", BigDecimal.valueOf(25.99), false),
-            new Product(6, "Ромашка", "Маленький белый цветок с желтой серединкой", BigDecimal.valueOf(5.99), true)
+            new Product(1, "Роза", "Красный цветок с шипами", 34.99, true),
+            new Product(2, "Тюльпан", "Весенний цветок различных оттенков", 20.99, true),
+            new Product(3, "Гвоздика", "Цветок с гофрированными лепестками", 20.99, false),
+            new Product(4, "Пионы", "Крупный цветок с пышными лепестками", 29.99, true),
+            new Product(5, "Лилии", "Элегантный цветок с сильным ароматом", 25.99, false),
+            new Product(6, "Ромашка", "Маленький белый цветок с желтой серединкой", 5.99, true)
     ));
 
     @Mock
@@ -32,7 +32,7 @@ class DefaultProductServiceTest {
 
     @Test
     public void shouldCatchNotFound_whenUpdatedUserById_success() {
-        ProductRequest product = new ProductRequest("Шиповник", "Кислая ягода", BigDecimal.valueOf(29.30), true);
+        ProductRequest product = new ProductRequest("Шиповник", "Кислая ягода", 29.30, true);
         doThrow(NotFoundException.class).when(productService).updateById(product, Integer.MAX_VALUE);
         assertThrows(NotFoundException.class, () -> productService.updateById(product, Integer.MAX_VALUE));
     }
